eliminate warnings generated during compiling

+++ src/solver/actnewton.cpp
@@ -59,7 +59,7 @@ namespace SAM {
     std::vector<double> stage_lambdas(d, 0);
     RegFunction *regfunc = new RegL1();
 
-    for (int i = 0; i < (int)lambdas.size(); i++) {
+    for (size_t i = 0; i < lambdas.size(); i++) {
       // start with the previous solution on the master path
       m_obj->set_model_param(model_master);
       m_obj->set_model_Xb(Xb_master);
